requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Label Matching
rapidfuzz>=3.0.0

# Progress & Logging
tqdm>=4.66.0
pyyaml>=6.0.0

# Jupyter Notebook
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
